+++ src/components/GoogleAnalytics.tsx
@@ -1,0 +1,24 @@
+import Script from "next/script";
+
+type Props = {
+  measurementId: string;
+};
+
+export default function GoogleAnalytics({ measurementId }: Props): React.ReactElement {
+  return (
+    <>
+      <Script
+        src={`https://www.googletagmanager.com/gtag/js?id=${measurementId}`}
+        strategy="afterInteractive"
+      />
+      <Script id="gtag-init" strategy="afterInteractive">
+        {`
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+          gtag('config', '${measurementId}');
+        `}
+      </Script>
+    </>
+  );
+}